requirements.txt
python-telegram-bot==13.15
environs==14.3.0
google-cloud-dialogflow==2.41.2